% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGeneInfo.R
\name{addGeneInfo}
\alias{addGeneInfo}
\title{Add Gene IDs to a proteinGroups Dataset}
\usage{
addGeneInfo(pg, uniprot)
}
\arguments{
\item{pg}{A data frame with protein IDs under the column "Protein.IDs". Needs to be cleaned with 'Fill_IDs'}

\item{uniprot}{A data frame with UniProt protein information, with protein IDs under the column "Entry". Also needs to contain}
}
\value{
A data frame with added gene IDs (Ensembl_ID, RefSeq_ID, Entrez_ID).
}
\description{
Merge a proteomics dataset with a UniProt database to add gene IDs (Ensembl, RefSeq, Entrez)
for the identified proteins.
}
\examples{
\dontrun{addGeneIDs(pg, uniprot)}

}
