% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkExperimentalDesign.R
\name{checkExperimentalDesign}
\alias{checkExperimentalDesign}
\title{Check Experimental Design}
\usage{
checkExperimentalDesign(ed, pg)
}
\arguments{
\item{ed}{A data frame containing the experimental design data.}

\item{pg}{A data frame containing the protein group data.}
}
\value{
Throws an error if the experimental design does not match the proteingroups format
}
\description{
This function checks the consistency and completeness of the experimental design data and protein group data.
}
\examples{
ed = createExperimentalDesign(pg)
checkExperimentalDesign(ed, pg)
}
