% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanPg.R
\name{cleanPg}
\alias{cleanPg}
\title{Clean protein groups data}
\usage{
cleanPg(pg, uniprot, species = "Hs")
}
\arguments{
\item{pg}{A data frame of protein groups, with one row per group and one or
more protein IDs in each group.}

\item{uniprot}{An object of class \code{uniprot} containing UniProt ID
mappings for the proteins in \code{pg}.}

\item{species}{A character string specifying the species to use for gene
information. The default is \code{'Hs'} (human).}
}
\value{
A data frame of cleaned protein groups.
}
\description{
This function takes a protein groups data frame \code{pg} and cleans it by
filling missing UniProt IDs using the \code{uniprot} object, removing
duplicate entries, and adding gene information. The resulting data frame is
returned.
}
\examples{
## Load protein groups and UniProt mappings
data(pg)
data(uniprot_hs)
## Clean protein groups
cleanPg(pg, uniprot_hs)

}
\seealso{
\code{\link{fill_IDs}}, \code{\link{removeDuplicates}},
\code{\link{addGeneInfo}}
}
