% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createExperimentalDesign.R
\name{createExperimentalDesign}
\alias{createExperimentalDesign}
\title{Create experimental design data from a proteinGroups file}
\usage{
createExperimentalDesign(pg)
}
\arguments{
\item{pg}{A proteomics dataset.}
}
\value{
A data frame with columns for label, condition, and replicate.
}
\description{
Given a proteomics dataset \code{pg}, this function extracts information about the experimental design
and returns a data frame with columns for sample label, condition, and replicate. The sample labels are
taken from the LFQ.intensity columns of the dataset.
}
\examples{
design <- createExperimentalDesign(pg)
}
