% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getImputationStats.R
\name{getImputationStats}
\alias{getImputationStats}
\title{Get statistics for padj and ratio values after multiple rounds of imputation.}
\usage{
getImputationStats(
  matrix_list,
  padj_threshold = 0.05,
  fc_threshold = 2,
  fraction_cutoff = 0.8
)
}
\arguments{
\item{matrix_list}{A list of matrices.}

\item{padj_threshold}{The p-value adjusted threshold (default: 0.05).}

\item{fc_threshold}{The fold change threshold (default: 2).}

\item{fraction_cutoff}{The fraction cutoff for significance (default: 0.8). In the default case, proteins need to be considered
significant in more than 80\% of the rounds of imputations}
}
\value{
A list containing the following components:
\describe{
\item{padj_fractions}{A matrix of adjusted p-value fractions.}
\item{ratio_fractions}{A matrix of ratio fractions.}
\item{results}{A data frame with statistics for padj, ratio, and significance values}
}

import matrixStats
}
\description{
Calculate the mean and standard deviation of p.adj values and ratios after multiple imputation.
In addition, it reports if proteins are significantly differentially expressed if both p.adj and ratio
meet the criteria set by padj_threshold and fc_threshold in more than fraction_cutoff * 100 \% of the imputations
}
