% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeMultiple.R
\name{imputeMultiple}
\alias{imputeMultiple}
\title{Impute Multiple}
\usage{
imputeMultiple(
  pg,
  expDesign,
  n,
  padj_threshold = 0.05,
  fc_threshold = 2,
  missing_threshold = 1
)
}
\arguments{
\item{pg}{A cleaned proteinGroups file}

\item{expDesign}{The experimental design matrix as produced by createExperimentalDesign (or similar)}

\item{n}{The number of iterations for imputation.}

\item{padj_threshold}{The p-value adjustment threshold (default: 0.05).}

\item{fc_threshold}{The fold change threshold (default: 2).}

\item{missing_threshold}{The missing value threshold (default: 1).}
}
\value{
A list containing n matrices with p.adjust values and ratios for each comparison.
}
\description{
Returns a list containing n matrices with p.adjust values and ratios for each pairwise comparison in experimentalDesign
}
