% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadExperimentalDesign.R
\name{loadExperimentalDesign}
\alias{loadExperimentalDesign}
\title{Load Experimental Design}
\usage{
loadExperimentalDesign()
}
\value{
A data frame containing the experimental design data.
}
\description{
This function loads the experimental design data from a user-selected file.
}
\examples{
\dontrun{
loadExperimentalDesign()
}


}
\keyword{data}
