% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadProteinGroups.R
\name{loadProteinGroups}
\alias{loadProteinGroups}
\title{Load Protein Groups}
\usage{
loadProteinGroups()
}
\value{
A data frame containing the protein group data.
}
\description{
This function loads protein group data from a user-selected file.
}
\examples{
\dontrun{loadProteinGroups()}

}
