% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performDEP.R
\name{performDEP}
\alias{performDEP}
\title{Perform Differential Expression Analysis using DEP}
\usage{
performDEP(pg, expDesign, thr = 1)
}
\arguments{
\item{pg}{A cleaned maxquant proteinGroups data frame containing LFQ protein expression data with gene names and protein IDs.}

\item{expDesign}{A data frame specifying the experimental design with at least two columns: 'Sample.ID' and 'Condition'.}

\item{thr}{A numeric value specifying the threshold for filtering out missing values (default is 1).}
}
\value{
A data frame containing gene/protein IDs and other identifiers with significant genes/proteins prioritized based on adjusted p-values and lfc.
}
\description{
This function performs differential expression analysis on protein expression data and prioritizes significant genes/proteins based on adjusted p-values and log fold change (lfc).
}
\examples{
\dontrun{performDEP(pg, expDesign)}
}
