% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performImputation.R
\name{performImputation}
\alias{performImputation}
\title{Perform Imputation on Intensity Data}
\usage{
performImputation(intensities, expDesign)
}
\arguments{
\item{intensities}{numeric matrix of intensity data with samples in columns and features in rows.}

\item{expDesign}{A data frame specifying the experimental design with at least two columns: 'Sample.ID' and 'Condition'.}
}
\value{
A numeric matrix with imputed intensity values for missing data.
}
\description{
This function takes log2-transformed and normalized LFQ intensity data as input and performs imputation to handle missing values.
Samples are split per conditions to determine whether MAR or MNAR is used. If all values for a condition are missing or in the lowest 10\% of intensity values
MNAR is applied. Otherwise MAR is applied. MAR is performed by MLE, MNAR by minProb method.
}
