% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performMultipleImputation.R
\name{performMultipleImputations}
\alias{performMultipleImputations}
\title{Perform Multiple Imputations}
\usage{
performMultipleImputations(
  pg,
  expDesign,
  n,
  padj_threshold = 0.05,
  fc_threshold = 2,
  missing_threshold = 1,
  fraction_cutoff = 0.8
)
}
\arguments{
\item{pg}{A cleaned maxquant proteingroups data frame}

\item{expDesign}{An experimental design created with createExperimentalDesign, or with similar structure.}

\item{n}{The number of imputations to perform.}

\item{padj_threshold}{The p-value adjusted threshold for significance (default: 0.05).}

\item{fc_threshold}{The fold change threshold for significance (default: 2).}

\item{missing_threshold}{The missing data threshold for imputation (default: 1).}

\item{fraction_cutoff}{The fraction cutoff (default: 0.8).}
}
\value{
The results of the imputation statistics.
}
\description{
Wrapper function that uses imputeMultiple and getImputationStats to
perform multiple imputations using the given parameters and returns the results of the imputations.
}
\seealso{
\code{\link{performImputation}}, \code{\link{imputeMultiple}},
\code{\link{getImputationStats}}
}
