% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pg}
\alias{pg}
\title{THRONCAT proteinGroups file}
\format{
\subsection{\code{pg}}{

A data frame with 1,350 rows and  columns:
}
}
\source{
\url{https://doi.org/10.1101/2022.03.29.486210}
}
\usage{
pg
}
\description{
An example maxquant proteinGroups file with 4x3 LFQ samples.
}
\keyword{datasets}
