% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeDuplicates.R
\name{removeDuplicates}
\alias{removeDuplicates}
\title{Remove Duplicates from a Protein Group DataFrame}
\usage{
removeDuplicates(pg)
}
\arguments{
\item{pg}{A protein group DataFrame containing Protein.IDs and Razor...unique.peptides columns.}
}
\value{
A list containing the protein group DataFrame with duplicates removed and a DataFrame of the removed duplicates.
}
\description{
This function removes duplicate entries from a given protein group DataFrame.
}
\examples{
pg <- data.frame(Protein.IDs = c("A", "A", "B", "C"), Razor...unique.peptides = c(1, 2, 3, 4))
removeDuplicates(pg)

}
