% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uniprot_hs}
\alias{uniprot_hs}
\title{uniprot_hs}
\format{
\subsection{\code{uniprot_hs}}{

A data frame with 208,318 rows and 11 columns:
}
}
\source{
\url{https://www.uniprot.org/}, \url{https://www.ensembl.org/info/data/biomart/index.html}
}
\usage{
uniprot_hs
}
\description{
UniProt database with all human entries.
}
\keyword{datasets}
